requirements.txt
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
duckdb>=1.0.0
python-dotenv>=1.0.1
tenacity>=8.4.0
groq>=0.9.0
httpx==0.27.2
